Break out status effects
